refactor: 디스커션 파일 상세 리팩터링
